Renamespace fixture loaders
