+++ src/Services/FixtureLoader/FixtureLoaderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\FixtureLoader;
 
 use Symfony\Component\Console\Output\OutputInterface;
