+++ src/Services/FixtureLoader/UserFixtureLoader.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\FixtureLoader;
 
 use App\Entity\User;
+use App\Services\UserDataProvider;
+use App\Services\UserService;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
